cleaned up services to check for network connectivity issues

+++ src/org/mythtv/service/dvr/RecordedCleanupService.java
@@ -18,19 +18,27 @@
  */
 package org.mythtv.service.dvr;
 
+import java.io.BufferedInputStream;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FilenameFilter;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.mythtv.service.MythtvService;
-import org.mythtv.service.dvr.cache.RecordedLruMemoryCache;
 import org.mythtv.service.util.UrlUtils;
 import org.mythtv.services.api.dvr.Program;
 import org.mythtv.services.api.dvr.Programs;
 
 import android.content.Intent;
 import android.util.Log;
+
+import com.fasterxml.jackson.core.JsonParseException;
+import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.joda.JodaModule;
 
 /**
  * @author Daniel Frey
@@ -46,12 +54,16 @@ public class RecordedCleanupService extends MythtvService {
     public static final String EXTRA_COMPLETE = "COMPLETE";
     public static final String EXTRA_COMPLETE_COUNT = "COMPLETE_COUNT";
 
-    private RecordedLruMemoryCache cache;
+    private final ObjectMapper mapper = new ObjectMapper();
+
+//    private RecordedLruMemoryCache cache;
 
     public RecordedCleanupService() {
 		super( "ProgamGuideDownloadService" );
 		
-		cache = new RecordedLruMemoryCache( this );
+//		cache = new RecordedLruMemoryCache( this );
+		mapper.registerModule( new JodaModule() );
+
 	}
 	
 	/* (non-Javadoc)
@@ -73,48 +85,63 @@ public class RecordedCleanupService extends MythtvService {
 	// internal helpers
 	
 	private void cleanup() {
-//		Log.v( TAG, "cleanup : enter" );
+		Log.v( TAG, "cleanup : enter" );
 		
 		int count = 0;
 		
 		File programCache = mFileHelper.getProgramDataDirectory();
 		if( null != programCache && programCache.exists() ) {
 
-			List<String> programGroups = new ArrayList<String>();
-			Programs programs = cache.get( RecordedDownloadService.RECORDED_FILE );
-			if( null != programs ) {
-				for( Program program : programs.getPrograms() ) {
+			File file = new File( programCache, RecordedDownloadService.RECORDED_FILE );
+			if( file.exists() ) {
+				try {
+					InputStream is = new BufferedInputStream( new FileInputStream( file ), 8192 );
+					Programs programs = mapper.readValue( is, Programs.class );
 
-					String title = UrlUtils.encodeUrl( program.getTitle() );
+					List<String> programGroups = new ArrayList<String>();
+					if( null != programs ) {
+						for( Program program : programs.getPrograms() ) {
 
-					if( !programGroups.contains( title ) ) {
-						programGroups.add( title );
-					}
-				}
+							String title = UrlUtils.encodeUrl( program.getTitle() );
 
-				FilenameFilter filter = new FilenameFilter() {
+							if( !programGroups.contains( title ) ) {
+								programGroups.add( title );
+							}
+						}
 
-					@Override
-					public boolean accept( File dir, String filename ) {
-						return filename.endsWith( ProgramGroupRecordedDownloadService.RECORDED_FILE );
-					}
+						FilenameFilter filter = new FilenameFilter() {
 
-				};
-				for( String filename : programCache.list( filter ) ) {
-//					Log.v( TAG, "cleanup : filename=" + filename );
+							@Override
+							public boolean accept( File dir, String filename ) {
+								return filename.endsWith( ProgramGroupRecordedDownloadService.RECORDED_FILE );
+							}
 
-					File deleted = new File( programCache, filename );
-					String programGroup = filename.substring( 0, filename.indexOf( ProgramGroupRecordedDownloadService.RECORDED_FILE ) );
-					if( !programGroups.contains( programGroup ) ) {
+						};
+						for( String filename : programCache.list( filter ) ) {
 
-						if( deleted.delete() ) {
-							count++;
+							File deleted = new File( programCache, filename );
+							String programGroup = filename.substring( 0, filename.indexOf( ProgramGroupRecordedDownloadService.RECORDED_FILE ) );
+							if( !programGroups.contains( programGroup ) ) {
 
-//							Log.v( TAG, "cleanup : deleted filename=" + filename );
+								if( deleted.delete() ) {
+									count++;
+
+									Log.v( TAG, "cleanup : deleted filename=" + filename );
+								}
+							}
 						}
 					}
+
+				} catch( JsonParseException e ) {
+					Log.e( TAG, "create : JsonParseException - error opening file 'recorded'", e );
+				} catch( JsonMappingException e ) {
+					Log.e( TAG, "create : JsonMappingException - error opening file 'recorded'", e );
+				} catch( IOException e ) {
+					Log.e( TAG, "create : IOException - error opening file 'recorded'", e );
 				}
+				
 			}
+			
 		}
 
 		Intent completeIntent = new Intent( ACTION_COMPLETE );
@@ -122,7 +149,7 @@ public class RecordedCleanupService extends MythtvService {
 		completeIntent.putExtra( EXTRA_COMPLETE_COUNT, count );
 		sendBroadcast( completeIntent );
 		
-//		Log.v( TAG, "cleanup : exit" );
+		Log.v( TAG, "cleanup : exit" );
 	}
 	
 }
